+++ srcs/common/Basic_process/extension/CPU/extend_constant_exec.h
@@ -44,19 +44,19 @@ namespace decx
 namespace bp {
 namespace CPUK {
     _THREAD_FUNCTION_ void _extend_constant1D_b32(const float* src, float* dst, const float val,
-        const decx::bp::extend_reflect_exec_params* b_rfct, const size_t _actual_w_v1, const size_t _original_w_v8);
+        const decx::bp::extend_reflect_exec_params* b_rfct, const uint64_t _actual_w_v1, const uint64_t _original_w_v8);
 
 
     _THREAD_FUNCTION_ void _extend_constant1D_b64(const double* src, double* dst, const double val,
-        const decx::bp::extend_reflect_exec_params* b_rfct, const size_t _actual_w_v1, const size_t _original_w_v4);
+        const decx::bp::extend_reflect_exec_params* b_rfct, const uint64_t _actual_w_v1, const uint64_t _original_w_v4);
 
 
     _THREAD_FUNCTION_ void _extend_constant1D_b8(const uint8_t* src, uint8_t* dst, const uint8_t val,
-        const decx::bp::extend_reflect_exec_params* b_rfct, const size_t _actual_w_v1, const size_t _original_w_v8);
+        const decx::bp::extend_reflect_exec_params* b_rfct, const uint64_t _actual_w_v1, const uint64_t _original_w_v8);
 
 
     _THREAD_FUNCTION_ void _extend_constant1D_b16(const uint16_t* src, uint16_t* dst, const uint16_t val,
-        const decx::bp::extend_reflect_exec_params* b_rfct, const size_t _actual_w_v1, const size_t _original_w_v8);
+        const decx::bp::extend_reflect_exec_params* b_rfct, const uint64_t _actual_w_v1, const uint64_t _original_w_v8);
 
 
     _THREAD_FUNCTION_ void _extend_H_constant2D_b32(const float* src, float* dst, const float val,
