+++ srcs/common/Basic_process/extension/CPU/x86/extend_reflect_exec.cxx
@@ -134,8 +134,8 @@ decx::bp::CPUK::_extend_reflect1D_b32(const float* __restrict   src,
                                       float* __restrict         buffer,
                                       float* __restrict         dst,
                                       const decx::bp::extend_reflect_exec_params* b_rfct,
-                                      const size_t              _actual_w_v1,
-                                      const size_t              _original_w_v8)
+                                      const uint64_t              _actual_w_v1,
+                                      const uint64_t              _original_w_v8)
 {
     const __m256i f = decx::bp::e_rfct_exep_get_shufflevar_f_b32(b_rfct);
     const __m256i b = decx::bp::e_rfct_exep_get_shufflevar_b_b32(b_rfct);
@@ -187,7 +187,7 @@ decx::bp::CPUK::_extend_H_reflect2D_b32(const float* __restrict   src,
 
     __m256 reg0, reg1, store;
 
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     for (uint32_t i = 0; i < _original_dims_v8.y; ++i) 
     {
@@ -233,7 +233,7 @@ decx::bp::CPUK::_extend_H_reflect2D_b32(const float* __restrict   src,
 
 
 _THREAD_FUNCTION_ void decx::bp::CPUK::_extend_reflect1D_b8(const uint8_t* src, uint8_t* buffer, uint8_t* dst,
-    const decx::bp::extend_reflect_exec_params* b_rfct, const size_t _actual_w_v1, const size_t _original_w_v8)
+    const decx::bp::extend_reflect_exec_params* b_rfct, const uint64_t _actual_w_v1, const uint64_t _original_w_v8)
 {
     const __m128i f = decx::bp::e_rfct_exep_get_shufflevar_f_b8(b_rfct);
     const __m128i b = decx::bp::e_rfct_exep_get_shufflevar_b_b8(b_rfct);
@@ -287,7 +287,7 @@ decx::bp::CPUK::_extend_H_reflect2D_b8(const uint8_t* __restrict   src,
 
     decx::utils::simd::xmm128_reg reg0, reg1, store;
 
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     for (uint32_t i = 0; i < _original_dims_v16.y; ++i) 
     {
@@ -368,7 +368,7 @@ decx::bp::CPUK::_extend_V_reflect2D_m256(float* __restrict   dst,        // poin
 
 
 _THREAD_FUNCTION_ void decx::bp::CPUK::_extend_reflect1D_b16(const uint16_t* src, uint16_t* buffer, uint16_t* dst,
-    const decx::bp::extend_reflect_exec_params* b_rfct, const size_t _actual_w_v1, const size_t _original_w_v8)
+    const decx::bp::extend_reflect_exec_params* b_rfct, const uint64_t _actual_w_v1, const uint64_t _original_w_v8)
 {
     const __m128i f = decx::bp::e_rfct_exep_get_shufflevar_f_b16(b_rfct);
     const __m128i b = decx::bp::e_rfct_exep_get_shufflevar_b_b16(b_rfct);
@@ -419,7 +419,7 @@ decx::bp::CPUK::_extend_H_reflect2D_b16(const uint16_t* __restrict   src,
 
     decx::utils::simd::xmm128_reg reg0, reg1, store;
 
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     for (uint32_t i = 0; i < _original_dims_v16.y; ++i) 
     {
@@ -466,8 +466,8 @@ decx::bp::CPUK::_extend_reflect1D_b64(const double* __restrict   src,
                                       double* __restrict         buffer,
                                       double* __restrict         dst,
                                       const decx::bp::extend_reflect_exec_params* b_rfct,
-                                      const size_t              _actual_w_v1,
-                                      const size_t              _original_w_v4)
+                                      const uint64_t              _actual_w_v1,
+                                      const uint64_t              _original_w_v4)
 {
     const __m256i f = decx::bp::e_rfct_exep_get_shufflevar_f_b64(b_rfct);
     const __m256i b = decx::bp::e_rfct_exep_get_shufflevar_b_b64(b_rfct);
@@ -524,7 +524,7 @@ decx::bp::CPUK::_extend_H_reflect2D_b64(const double* __restrict   src,
 
     decx::utils::simd::xmm256_reg reg0, reg1, store;
 
-    size_t dex_src = 0, dex_dst = 0;
+    uint64_t dex_src = 0, dex_dst = 0;
 
     for (uint32_t i = 0; i < _original_dims_v8.y; ++i) 
     {
